Website/requirements.txt
streamlit
requests
numpy
scikit-image
opencv-python
Pillow
opencv-contrib-python
